requirements.txt
tensorflow>=2.12
keras>=2.12
numpy
pillow
matplotlib
scikit-learn
 
